fix default args type

+++ arguments.py
@@ -70,7 +70,7 @@ def get_args():
     parser.add_argument('--hindsight_edge', type=bool, default=True, help='use hindsight edges')
     parser.add_argument('--edge_prior', type=float, default=0.5, help='default value for edges')
     parser.add_argument('--unordered_edge', type=bool, default=True, help='if the agent learns unordered_edge SR')
-    parser.add_argument('--epsilon_edge_exploration', type=int, default=0, help='at step of a path to the frontier, the agent has an epsilon chance to take a random edge')
+    parser.add_argument('--epsilon_edge_exploration', type=float, default=0, help='at step of a path to the frontier, the agent has an epsilon chance to take a random edge')
     
     # TODO : Add sampling for agent paths.  
     #parser.add_argument('--sample_path', type=bool, default=False, help='if the agent takes the best path or sample it')
